Adapting Spark3.3
